AIX FFM port (phase 1): move non-JNA classes to oshi-common
Relocates AIX classes that don't touch JNA out of oshi-core and into
oshi-common, mirroring the structure used by the Solaris and OpenBSD
ports. This is a pure move with no behavior changes; subsequent commits
will add the FFM bindings and split the JNA-using classes.

* Drivers Ls/Lscfg/Lspv/Lssrad/Uptime/Who move to
  oshi.driver.common.unix.aix (the .common segment avoids a split
  package with the JNA-only PsInfo / perfstat drivers that remain in
  oshi-core).
* HAL classes AixBaseboard/ComputerSystem/Firmware/GraphicsCard/
  PowerSource/Sensors/SoundCard/UsbDevice move to
  oshi.hardware.common.platform.unix.aix; package-private classes are
  raised to public so the JNA HAL in oshi-core can still construct
  them across module boundaries.
* OS classes AixFileSystem/InstalledApps/OSFileStore move to
  oshi.software.common.os.unix.aix.
* Three driver tests follow their drivers; the AIX-only @EnabledOnOs
  guards remain in place.
* module-info.java exports the three new common packages and
  module-info.test opens them to org.junit.platform.commons.
* New package-info.java files replace the moved ones for both the
  oshi-core JNA-only HAL package and the new oshi-common OS package.
