+++ oshi-common/src/main/java/oshi/driver/common/unix/aix/Ls.java
@@ -1,8 +1,8 @@
 /*
- * Copyright 2020-2022 The OSHI Project Contributors
+ * Copyright 2020-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.driver.unix.aix;
+package oshi.driver.common.unix.aix;
 
 import java.util.HashMap;
 import java.util.Map;
