+++ oshi-common/src/main/java/oshi/hardware/common/platform/unix/aix/AixFirmware.java
@@ -1,8 +1,8 @@
 /*
- * Copyright 2020-2022 The OSHI Project Contributors
+ * Copyright 2020-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.hardware.platform.unix.aix;
+package oshi.hardware.common.platform.unix.aix;
 
 import oshi.annotation.concurrent.Immutable;
 import oshi.hardware.common.AbstractFirmware;
@@ -11,7 +11,7 @@ import oshi.hardware.common.AbstractFirmware;
  * Firmware data.
  */
 @Immutable
-final class AixFirmware extends AbstractFirmware {
+public final class AixFirmware extends AbstractFirmware {
 
     private final String manufacturer;
     private final String name;
