+++ oshi-common/src/main/java/oshi/hardware/common/platform/unix/aix/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) 2020-2022 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Provides information about hardware such as Memory, Power Sources, and Processor on AIX systems
+ */
+package oshi.hardware.common.platform.unix.aix;
