+++ oshi-common/src/main/java/oshi/software/common/os/unix/aix/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Provides abstract base classes and pure-Java implementations for AIX software, shared by the JNA and FFM modules.
+ */
+package oshi.software.common.os.unix.aix;
