+++ oshi-common/src/test/java/module-info.test
@@ -64,6 +64,12 @@
   com.github.oshi.common/oshi.software.common.os.unix.openbsd=org.junit.platform.commons
 --add-opens
   com.github.oshi.common/oshi.util.common.platform.unix.openbsd=org.junit.platform.commons
+--add-opens
+  com.github.oshi.common/oshi.driver.common.unix.aix=org.junit.platform.commons
+--add-opens
+  com.github.oshi.common/oshi.hardware.common.platform.unix.aix=org.junit.platform.commons
+--add-opens
+  com.github.oshi.common/oshi.software.common.os.unix.aix=org.junit.platform.commons
 
 --add-modules
   org.hamcrest,org.slf4j.simple
