+++ oshi-common/src/test/java/oshi/driver/common/unix/aix/LssradTest.java
@@ -1,8 +1,8 @@
 /*
- * Copyright 2022 The OSHI Project Contributors
+ * Copyright 2022-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.driver.unix.aix;
+package oshi.driver.common.unix.aix;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.anEmptyMap;
