+++ oshi-core/src/main/java/oshi/hardware/platform/unix/aix/AixCentralProcessor.java
@@ -20,7 +20,7 @@ import com.sun.jna.platform.unix.aix.Perfstat.perfstat_cpu_total_t;
 import com.sun.jna.platform.unix.aix.Perfstat.perfstat_partition_config_t;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.driver.unix.aix.Lssrad;
+import oshi.driver.common.unix.aix.Lssrad;
 import oshi.driver.unix.aix.perfstat.PerfstatConfig;
 import oshi.driver.unix.aix.perfstat.PerfstatCpu;
 import oshi.hardware.CentralProcessor.ProcessorCache.Type;
