+++ oshi-core/src/main/java/oshi/hardware/platform/unix/aix/package-info.java
@@ -1,8 +1,8 @@
 /*
- * Copyright (c) 2020-2022 The OSHI Project Contributors
+ * Copyright 2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
 /**
- * Provides information about hardware such as Memory, Power Sources, and Processor on AIX systems
+ * Provides JNA-backed hardware implementations for AIX.
  */
 package oshi.hardware.platform.unix.aix;
