+++ oshi-core/src/main/java/oshi/software/os/unix/aix/AixOperatingSystem.java
@@ -26,12 +26,14 @@ import com.sun.jna.platform.unix.aix.Perfstat.perfstat_partition_config_t;
 import com.sun.jna.platform.unix.aix.Perfstat.perfstat_process_t;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.driver.unix.aix.Uptime;
-import oshi.driver.unix.aix.Who;
+import oshi.driver.common.unix.aix.Uptime;
+import oshi.driver.common.unix.aix.Who;
 import oshi.driver.unix.aix.perfstat.PerfstatConfig;
 import oshi.driver.unix.aix.perfstat.PerfstatProcess;
 import oshi.jna.platform.unix.AixLibc;
 import oshi.software.common.AbstractOperatingSystem;
+import oshi.software.common.os.unix.aix.AixFileSystem;
+import oshi.software.common.os.unix.aix.AixInstalledApps;
 import oshi.software.os.ApplicationInfo;
 import oshi.software.os.FileSystem;
 import oshi.software.os.InternetProtocolStats;
